+++ builder/src/main/java/com/stratio/cassandra/lucene/builder/Builder.java
@@ -488,10 +488,19 @@ public abstract class Builder {
     /**
      * Returns a new {@link GeoTransformation.Centroid}.
      *
-     * @return a new buffer transformation
+     * @return a new centroid transformation
      */
     public static GeoTransformation.Centroid centroidGeoTransformation() {
         return new GeoTransformation.Centroid();
+    }
+
+    /**
+     * Returns a new {@link GeoTransformation.Centroid}.
+     *
+     * @return a new convex hull transformation
+     */
+    public static GeoTransformation.ConvexHull convexHullGeoTransformation() {
+        return new GeoTransformation.ConvexHull();
     }
 
     /**