+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix invalid WKT shapes by zero-buffering them
 * Fix memory consumption issues with high fetch sizes
 * Early terminate in inner pagination (improves allow filtering performance)
+* Add transformation for getting the convex hull of a geographical shape
 
 ## 3.0.5.0 (22 April 2016)